requirements.txt
huggingface_hub>=0.27.0
openai>=1.40.0
pandas>=2.0.0
streamlit>=1.39.0
pyyaml>=6.0
python-dotenv>=1.0.0
